DEPI-PROJECT-API_CHAT_RAG_HISTORY_TEMPLATES_br/DEPI-PROJECT-API_CHAT_HISTORY_TEMPLATES_br/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-dotenv==1.1.1
pydantic-settings==2.11.0
SQLAlchemy==2.0.44
aiosqlite==0.21.0

# LLM Providers
openai>=1.0.0
cohere>=4.0.0
langchain>=0.1.0
langchain-core>=0.1.0

# RAG Components
chromadb>=0.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6